GFORMS-1602 - Choice indexes/values in robotics need to handle versioning

+++ app/uk/gov/hmrc/gform/validation/ValidationUtil.scala
@@ -148,7 +148,7 @@ object ValidationUtil {
                 case Some(errors) => FieldError(formComponent, currentValue, errors)
                 case None         => FieldOk(formComponent, currentValue)
               }
-              HtmlFieldId.indexed(formComponent.id, index.toInt) -> formFieldValidationResult
+              HtmlFieldId.indexed(formComponent.id, index) -> formFieldValidationResult
             }.toMap
             ComponentField(formComponent, optionalData)
           }
